Issue #150: Minor changes: sort scores data in case-insensitive mode.
